requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
awscli
blinker
boto3
botocore
cachetools
certifi
charset-normalizer
click
colorama
dataclasses-json
docutils
exceptiongroup
frozenlist
gitdb
GitPython
idna
importlib-metadata
Jinja2
jmespath
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
numpy
opensearch-py
packaging
pandas
Pillow
pip
protobuf
pyarrow
pyasn1
pydantic
pydantic_core
pydeck
Pygments
pypdf
python-dateutil
python-dotenv
pytz
PyYAML
referencing
requests
rich
rpds-py
rsa
s3transfer
setuptools
six
smmap
sniffio
SQLAlchemy
streamlit
tenacity
toml
toolz
tornado
typing_extensions
typing-inspect
tzdata
tzlocal
urllib3
validators
wheel
yarl
zipp